Caching and using the API directly
This is getting much closer to where I want to be. I like that the happy path reads as straight line code and that errors are handled out of band in the catch block. I think this will let me remove all of the previous API code since the tumblr api does most of that work now.

I have some future work to do to extract the error handling into middleware that can be used by all of the tubmlr reuqests.

I also have work to do to remove the previous caching idea. This version caches out to files which means the server can go down completely and when it comes back everything is the same as it was.

Eventually I'm going to need to address the `buildRSSFeed` function to understand what's going on and make sure it's what I really want. `sigh` I'm afraid of that function. It mostly does a really good job. The feature I want to add most is the ability to proxy the images through another host.